GTACord/requirements.txt
py-cord==2.0.0b1
aiosqlite
python-dotenv